Extracted the Delete keys component into a separate file
