+++ packages/www/components/Dashboard/SigningKeysTable/index.tsx
@@ -1,18 +1,5 @@
-import {
-  Text,
-  Heading,
-  Box,
-  AlertDialog,
-  AlertDialogAction,
-  AlertDialogCancel,
-  AlertDialogContent,
-  AlertDialogDescription,
-  AlertDialogTitle,
-  Button,
-  Flex,
-  useSnackbar,
-} from "@livepeer/design-system";
-import { useCallback, useMemo, useState } from "react";
+import { Heading, Box } from "@livepeer/design-system";
+import { useCallback, useMemo } from "react";
 import { useApi } from "../../../hooks";
 import Table, { Fetcher, useTableState } from "components/Dashboard/Table";
 import { Cross1Icon, PlusIcon } from "@radix-ui/react-icons";
@@ -24,7 +11,7 @@ import {
   SigningKeysTableData,
 } from "./helpers";
 import CreateKeyDialog from "./CreateKeyDialog";
-import Spinner from "../Spinner";
+import DeleteKeysDialog from "./DeleteKeysDialog";
 
 const SigningKeysTable = ({
   title = "Signing Keys",
@@ -35,15 +22,13 @@ const SigningKeysTable = ({
   pageSize?: number;
   tableId: string;
 }) => {
-  const { getSigningKeys, deleteSigningKey } = useApi();
+  const { getSigningKeys } = useApi();
   const { state, stateSetter } = useTableState<SigningKeysTableData>({
     pageSize,
     tableId,
   });
   const deleteDialogState = useToggleState();
   const createDialogState = useToggleState();
-  const [openSnackbar] = useSnackbar();
-  const [savingDeleteDialog, setSavingDeleteDialog] = useState(false);
 
   const columns = useMemo(makeColumns, []);
 
@@ -51,26 +36,6 @@ const SigningKeysTable = ({
     async (state) => rowsPageFromState(state, getSigningKeys),
     []
   );
-
-  const onDeleteSigningKeys = useCallback(async () => {
-    setSavingDeleteDialog(true);
-    await Promise.all(
-      state.selectedRows.map((row) => deleteSigningKey(row.id))
-    );
-    openSnackbar(
-      `${state.selectedRows.length} signing key${
-        state.selectedRows.length > 1 ? "s" : ""
-      } deleted.`
-    );
-    setSavingDeleteDialog(false);
-    await state.invalidate();
-    deleteDialogState.onOff();
-  }, [
-    deleteSigningKey,
-    deleteDialogState.onOff,
-    state.selectedRows.length,
-    state.invalidate,
-  ]);
 
   return (
     <>
@@ -113,56 +78,7 @@ const SigningKeysTable = ({
       />
 
       {/* Delete dialog */}
-      <AlertDialog
-        open={deleteDialogState.on}
-        onOpenChange={deleteDialogState.onOff}>
-        <AlertDialogContent
-          css={{ maxWidth: 450, px: "$5", pt: "$4", pb: "$4" }}>
-          <AlertDialogTitle asChild>
-            <Heading size="1">
-              Delete {state.selectedRows.length} signing key
-              {state.selectedRows.length > 1 && "s"}?
-            </Heading>
-          </AlertDialogTitle>
-          <AlertDialogDescription asChild>
-            <Text
-              size="3"
-              variant="gray"
-              css={{ mt: "$2", lineHeight: "22px" }}>
-              This will permanently remove the signing key
-              {state.selectedRows.length > 1 && "s"}. This action cannot be
-              undone.
-            </Text>
-          </AlertDialogDescription>
-
-          <Flex css={{ jc: "flex-end", gap: "$3", mt: "$5" }}>
-            <AlertDialogCancel asChild>
-              <Button size="2" onClick={deleteDialogState.onOff} ghost>
-                Cancel
-              </Button>
-            </AlertDialogCancel>
-            <AlertDialogAction asChild>
-              <Button
-                size="2"
-                disabled={savingDeleteDialog}
-                onClick={onDeleteSigningKeys}
-                variant="red">
-                {savingDeleteDialog && (
-                  <Spinner
-                    css={{
-                      color: "$hiContrast",
-                      width: 16,
-                      height: 16,
-                      mr: "$2",
-                    }}
-                  />
-                )}
-                Delete
-              </Button>
-            </AlertDialogAction>
-          </Flex>
-        </AlertDialogContent>
-      </AlertDialog>
+      <DeleteKeysDialog state={state} deleteDialogState={deleteDialogState} />
 
       {/* Create dialog */}
       <CreateKeyDialog
